Modal: Update buttons & title
Right aligns the modal buttons.
Changes the modal buttons top margin
Hides the cancel button when a respective method is not provided.
Changes the tilte based on the styleguide.
Adds a titleDetails property based on the styleguide.
Adds extra modal stories and makes most of them to have the modals open by default.

Resolves #194
Change-Type: minor