docs(Stories): Use Provider component in individual stories
Fixes #516

Previously the `Provider` component was added as a decorator in a single
place (`.storybook/config.js`), this was great for reducing code
repetition, but made it difficult for newcomers to reason about how to
use components. The `Provider` component is now loaded into each
seperate story file, allowing the stories to work as complete examples
of how to use rendition.
This change also removes the `addon-info` storybook package, which
doesn't work well with styled-components and has been superseded by our
improved documentation.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>