Remove additionalProperties from Form schemas
As of react-jsonschema-form v1.0.6, if the `additionalProperties` keyword is a
truthy value, additional UI controls will be added so that users can
freely add additional properties. This is undesired behaviour in
rendition, so the schema value passed to the underlying Form component
now has `additionalProperties` forcibly set to false.

Change-type: patch
Signed-off-by: Lucian <lucian.buzzo@gmail.com>