Form: Add alpha support for patternProperties
This change adds support to the Form component for the JSON schema
"patternProperties" keyword. This is an early implementation and needs
further work to be production ready.
I also took this opportunity to tidy up JSDOM setup, by defining the
document global and other useful values in one place.

Change-type: minor
Signed-off-by: Lucian <lucian.buzzo@gmail.com>