Fixes #759 - a problem where if a formula definition would be inside of an array, it would not get evaluated properly.
More info:
We preprocess the data before sending it to temen to evaluate - we put the formulas from the schema in the relevant places inside the data. This was done for objects but not for arrays of objects. The code in this PR basically goes through the data object and makes sure that for every item in every array on any level we invoke the substitution of existing data for the formula.

In the future - it may be good to move all this `data + schema => data_with_formulas` logic inside reconfix maybe, so that JellyForm only calls these functions ? Or maybe there is an existing functionality that we could reuse here ?

Change-type: patch