Merge Pill component into Badge component
Connects to #801

Since the visual change to the Button component makes it look very
similar to the Pill component, the Pill component has had its rounded
corners removed. This makes the Pill component essentially a small
Badge, so the decision was made to simply merge the two components
together. The equivalent of the Pill component can now be created using
the `small` and `xsmall` properties on the Badge component.
The Badge component has also been modified to use children as the text
content instead of a `text` property.

Change-type: major
Signed-off-by: Lucian <lucian.buzzo@gmail.com>