filters: Fix default datepicker date
The default Date that the datepicker input would receive wasn't in the right format. As a result it would fail to set the current datetime on first render. Any consequent props that the component would receive would retrigger this conditional resulting in an never ending loop.

If the user would try to select a date while a polling is in place, the field would reset. This fix ensures that the initial value is correct.

Fixes: #887
Change-type: patch
Signed-off-by: Dimitrios Lytras <dimitrios@balena.io>