feat: Add support for passing extra context to JsonSchemaRenderer
This extraContext prop is used when evaluating UI schema using json-e.

Additionally, json-e is now run (recursively) just on the UI schema for primitive values. This means we can now use json-e to evaluate the values of array items in the UI schema (see the 'A List Widget' example).

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>