Move JsonSchemaRenderer to core components and remove default widgets/formats
This is a breaking change that will require an action. Firstly, if you were
using JsonSchemaRenderer from `extra`, you now need to rename JsonSchemaRenderer
to Renderer, and import it from 'rendition'. The next thing to do is pass
all extra formats that you want the renderer to support that are not default
(namely Markdown and Mermaid) to the Provider's `widgets` prop. Also, the
error boundary as part of the renderer was removed and you need to add that
by yourself if needed.

Change-type: major
Signed-off-by: Stevche Radevski <stevche@balena.io>