Change how extra widgets are registered on Form
This is a breaking change and requires some action if you
are using extra widgets. You need to remove all `registerWidget`
calls and instead pass the extra widgets you want to use to either
the Form as `extraFormats` or set them on the Provider's `widgets` prop

Moreover, if you used a reference to call `change` and `submit`
you will only be able to call submit now by passing the ref directly to the form
and change was completely removed.

Change-type: major
Signed-off-by: Stevche Radevski <stevche@balena.io>