DropDownButton: Hide empty items
React doesn't offer a way to detect and not render
such items at all any more. This is a convenient
improvement though since it allows better
encapsulation & more flexibility in cases that that
the items need to be auto-generated based from
a definition list (eg: for device actions).

Change-type: minor
Signed-off-by: Thodoris Greasidis <thodoris@balena.io>