DropDownButton: add support for dropping 'up'
When the DropDownButton is located at the bottom of the screen, it can be useful to specify that the DropDownButton's menu should open _above_ the button, rather than below (as is the default).

Obviously it would be even nicer to be able to specify that it should automatically work out whether to open up or down - but adding this `dropUp` prop is a good first step for now.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>