Renderer: handle anyOf options
The Renderer component now checks if `anyOf` options are specified in the schema. If they are, it uses `skhema.match` to find the option that most closely matches the value and returns that schema.

This also works with array items that can be of different types. See examples for details.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>