Remove deprecated recompose package
Recompose is no longer being actively maintained and since we are using
a very small subset of its API its a simple thing to replace.
The `getDisplayName()` method is trivial and only used in a single
place.
The `compose()` method is functionally the same as `lodash.flowRight()`
so can be swapped out.
This leaves `withProps()` which again is relatively straightforward.
I've added a new utils function that handles the props wrapping this
method previously provided.

Connects-to: #1361
Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>